Move (app)-only modal CSS imports off store.ts; trim Sentry tracing
Two related bundle-size fixes:

1) Split modal show* helpers into core (store.ts) and (app)-only (app.ts)

PSI flagged a ~16 KB "100% unused" premium CSS chunk on /blog. The
trigger was lib/modal/store.ts: it statically imported CSS modules
from PremiumUpgradeModal, SubscriptionCheckoutModal, PaymentProcessing,
WelcomeToPremium, Welcome, WalkthroughConfirm and AvatarEdit, because
each show* helper needs to forward the modal's hashed overlay/wrapper
class names to BaseModal. Since store.ts is loaded everywhere via
useModalStore (mounted by GlobalModalProvider in the root layout),
every blog/articles/landing/marketing route was bundling those CSS
modules even though the modals never render there.

Move the (app)-only helpers (and their CSS imports) to a new
lib/modal/app.ts. store.ts stays pure state + the always-safe helpers
(showModal, hideModal, useModalStore, showConfirmation, showInfo) and
no longer pulls any (app)-only modal CSS. 25 callsites updated to
import from "@/lib/modal/app" instead of "@/lib/modal" (or
"@/lib/modal/store") where they reach for an (app)-only helper.

The premium CSS chunk should drop off non-(app) routes after deploy.

2) Sentry: removeTracing: true

Add removeTracing to the Sentry webpack treeshake config so the
BrowserTracing/performance SDK is stripped from the production bundle.
We capture errors only — no tracing data is consumed — so this is
safe and saves ~30–40 KB off the shared Sentry chunk on every route.

Note: this PR does NOT address the separate ~17 KB CodingExercise CSS
chunk PSI also flagged on /blog. That chunk is a Webpack CSS-grouping
quirk (28 CSS Modules co-extracted because they share split-chunks
groups; ExternalFooter on the blog route is enough to pull the whole
chunk in). Needs a focused next.config.ts splitChunks tweak —
separate follow-up.
