+++ app/app/(app)/projects/PremiumProjectCard.tsx
@@ -1,7 +1,7 @@
 import { type ProjectData } from "@/lib/api/projects";
 import { ProjectIcon } from "@/components/icons/ProjectIcon";
 import LockIcon from "@/icons/lock.svg";
-import { showPremiumUpgradeModal } from "@/lib/modal";
+import { showPremiumUpgradeModal } from "@/lib/modal/app";
 import styles from "./ProjectCard.module.css";
 
 interface PremiumProjectCardProps {
