+++ app/lib/modal/index.ts
@@ -1,21 +1,8 @@
+// Public modal API for callsites outside (app).
+// See ./app.ts for the (app)-only show* helpers.
+
 // Export the provider component
 export { GlobalModalProvider } from "./GlobalModalProvider";
 
-// Export the modal functions for global usage
-export {
-  hideModal,
-  showConfirmation,
-  showInfo,
-  showModal,
-  showSubscriptionCheckout,
-  showSubscriptionModal,
-  showSubscriptionSuccess,
-  showPaymentProcessing,
-  showPaymentConfirming,
-  showPaymentVerificationFailed,
-  showPremiumUpgradeModal,
-  showWelcomeToPremium,
-  showAvatarEditModal,
-  showWelcomeModal,
-  useModalStore
-} from "./store";
+// Core modal trigger helpers — safe to import from any route.
+export { hideModal, showConfirmation, showInfo, showModal, useModalStore } from "./store";
